The most significant changes in the code include the removal of the `setPosition` method from the `Map.h` file, the change of window creation mode to fullscreen in the `Globals.cpp` file, and the modification of the `grid` method call in the `GlobalMap_PG.h` file to include the position of the chunk as a parameter. Additionally, the `distanceBetweenPoints` method in the `GlobalValues_PG.h` file was changed to return a float, and the debug mode was turned off in the `World_PG.h` file.
1. The `setPosition` method was removed from the `Map.h` file. This method was used to set the position of the map. (Map.h)
2. The window creation mode was changed from `sf::Style::Default` to `sf::Style::Fullscreen` in the `Globals.cpp` file. This change will make the window appear in fullscreen mode instead of the default windowed mode. (Globals.cpp)
3. The `grid` method call in the `GlobalMap_PG.h` file was modified to include the position of the chunk as a parameter. This change was made because the `setPosition` method was removed. (GlobalMap_PG.h)
4. The `distanceBetweenPoints` method in the `GlobalValues_PG.h` file was changed to return a float. This was done by adding a static cast to the return value. (GlobalValues_PG.h)
5. The `debug` variable was set to `false` in the `World_PG.h` file. This means that the debug mode is turned off by default (why tf was it on by default???). (World_PG.h)

+++ PirateGame/Header/GlobalValues_PG.h
@@ -34,7 +34,7 @@ namespace PirateGame {
 		}
 
 		float distanceBetweenPoints(const sf::Vector2f& point1, const sf::Vector2f& point2) {
-			return sqrt(pow(point1.x - point2.x, 2) + pow(point1.y - point2.y, 2));
+			return static_cast<float>(sqrt(pow(point1.x - point2.x, 2) + pow(point1.y - point2.y, 2)));
 		}
 
 		sf::Vector2f limitVector(const sf::Vector2f& vector, float max) {
